5.11

+++ src/main/java/com/zllUserCenter/findfriendbackend/model/domain/UserTeam.java
@@ -22,7 +22,7 @@ public class UserTeam {
     /**
      * 用户id
      */
-    private Integer userId;
+    private long userId;
 
     /**
      * 队伍id
